src/requirements.txt
functions-framework==3.8.1
google-auth==2.38.0
google-api-python-client==2.166.0
google-cloud-logging==3.11.4
requests==2.32.3